feat(ScreenSpinner): fix styleguide component name
