+++ packages/vkui/src/components/ScreenSpinner/types.tsx
@@ -1,0 +1,7 @@
+import type { SpinnerProps } from '../Spinner/Spinner';
+
+export interface ScreenSpinnerProps extends SpinnerProps {
+  state?: 'loading' | 'cancelable' | 'done' | 'error';
+  mode?: 'shadow' | 'overlay';
+  cancelLabel?: string;
+}
