+++ packages/vkui/src/index.ts
@@ -124,11 +124,10 @@ export type {
 export { ActionSheetItem } from './components/ActionSheetItem/ActionSheetItem';
 export type { ActionSheetItemProps } from './components/ActionSheetItem/ActionSheetItem';
 export { ActionSheetDefaultIosCloseItem } from './components/ActionSheet/ActionSheetDefaultIosCloseItem';
-export { ScreenSpinner, ScreenSpinnerContext } from './components/ScreenSpinner/ScreenSpinner';
-export type {
-  ScreenSpinnerProps,
-  ScreenSpinnerContextProps,
-} from './components/ScreenSpinner/ScreenSpinner';
+export { ScreenSpinner } from './components/ScreenSpinner/ScreenSpinner';
+export { ScreenSpinnerContext } from './components/ScreenSpinner/context';
+export type { ScreenSpinnerProps } from './components/ScreenSpinner/ScreenSpinner';
+export type { ScreenSpinnerContextProps } from './components/ScreenSpinner/context';
 export { Snackbar } from './components/Snackbar/Snackbar';
 export type { SnackbarProps } from './components/Snackbar/Snackbar';
 export { Tooltip } from './components/Tooltip/Tooltip';
